+++ counterpartylib/lib/transaction.py
@@ -325,6 +325,7 @@ def construct (db, tx_info, encoding='auto',
 
     global UTXO_LOCKS
 
+    desired_encoding = encoding
     (source, destination_outputs, data) = tx_info
 
     if dust_return_pubkey:
@@ -480,6 +481,11 @@ def construct (db, tx_info, encoding='auto',
     change_quantity = 0
     sufficient_funds = False
     final_fee = fee_per_kb
+    desired_input_count = 1
+
+    if encoding == 'multisig' and data_array and util.enabled('bytespersigop'):
+        desired_input_count = len(data_array) * 2
+
     for coin in use_inputs:
         logger.debug('New input: {}'.format(print_coin(coin)))
         inputs.append(coin)
@@ -502,7 +508,8 @@ def construct (db, tx_info, encoding='auto',
         # If change is necessary, must not be a dust output.
         if change_quantity == 0 or change_quantity >= regular_dust_size:
             sufficient_funds = True
-            break
+            if len(inputs) >= desired_input_count:
+                break
 
     if not sufficient_funds:
         # Approximate needed change, fee by with most recently calculated
@@ -536,6 +543,23 @@ def construct (db, tx_info, encoding='auto',
     else:
         change_output = None
 
+    # in bitcoin core v0.12.1 a -bytespersigop was added that messes with bare multisig transactions,
+    #  as a safeguard we fall back to pubkeyhash encoding when unsure
+    # when len(inputs) > len(data_outputs) there's more bytes:sigops ratio and we can safely continue
+    if encoding == 'multisig' and inputs and data_output and len(inputs) < len(data_array) * 2 and util.enabled('bytespersigop'):
+        # if auto encoding we can do pubkeyhash encoding instead
+        if desired_encoding == 'auto':
+            return construct(db, tx_info,
+                             encoding='pubkeyhash',
+                             fee_per_kb=fee_per_kb,
+                             regular_dust_size=regular_dust_size,
+                             multisig_dust_size=multisig_dust_size,
+                             op_return_value=op_return_value,
+                             exact_fee=exact_fee, fee_provided=fee_provided, provided_pubkeys=provided_pubkeys,
+                             allow_unconfirmed_inputs=allow_unconfirmed_inputs, unspent_tx_hash=unspent_tx_hash, custom_inputs=custom_inputs)
+        # otherwise raise exception
+        else:
+            raise exceptions.EncodingError("multisig will be rejected by Bitcoin Core >= v0.12.1, you should use `encoding=auto` or `encoding=pubkeyhash`")
 
     # Serialise inputs and outputs.
     unsigned_tx = serialise(encoding, inputs, destination_outputs,